requirements.txt
scikit-learn==0.19.1
numpy==1.14.5
scipy==1.1.0
tensorflow-gpu==1.8.0
jupyter==1.0.0
matplotlib==2.2.2
sympy==1.1.1
requests==2.19.1
nltk==3.3
pandas==0.23.1
gensim==3.4.0
